Unpack all coordinate points in one go, 5-9x speedup
When unpacking the list of coordinate points, each point pair is
unpacked one at a time, making it very slow for long lists of points. So
a fairly simple change to read all points at once, which btw is in line
with the existing convention elsewhere, leads to 5-9x speedup when
reading shapes, with bigger gains for files with more complex shapes
(more points).