More flexible with shapefile filetypes, in line with the docs
Fixes #35.

The docs for the reader say that no exception is thrown if any of the filetypes are missing until an operation is called that depends on that file, and that the .shx file is optional. This was only the case when giving file-like objects for each filetype, not when specifying the filename. And missing .shx files raised errors. 

This commit allows the same flexibility when specifying filenames, with the added benefit of reading standalone dbf files.