Stop zero from sneaking into mbox calculation
Changed the mbox calculation to use same structure as zbox, while preserving the intent of setting mbox to zeros in case of no m values. 

In the previous implementation the list of m values was instantiated with a zero, presumably to set an mbox of zeros in case of no m values, but this would eg force the mbox to be 0-30 in a case where the true range was 15-30. Unless there is something I don't know about m values, this should not be the case.