Fix error when writing file z/mbox.
Previously, the code for updating the file z or m bbox, seems to have been based on the normal bbox, and assumed four bbox values, resulting in IndexError. It should only be two (m/zmin, m/zmax). Fixes #128.

Also, if any of the coordinates in z or m type shapefiles are missing their z or m values, previously would silently skip and break on the first error found. Now only skips those that are missing so can get the best box possible, and raises a warning to notify the user.