Initiate bbox as None, rather than as 0s which can lead to problems
When writing a new shapefile, initialize bbox with None rather than assume 0s. Bbox with 0s has created problems and has been raised in issues in the past.
If bbox is still None by the time of saving, means the shapefile contains no valid geometries, and then we can set the bbox to 0s.
Not clear that this is the right way to write missing bbox. ESRI spec says the bbox values are 'unspecified', so not sure what that means.
But at least setting to 0s should be safe as is the same behavior as in previous versions, and makes sense since the spec says Z and M bounds should be set to 0 for non-Z/M type shapefiles.