Removed incorrect interpretation of 0s in file mbox
The ESRI shapefile spec says the file mbox is set to 0.0 when the file is not of type M or Z. I believe this part of the spec was incorrectly interpreted in previous versions, so I removed a part that incorrectly skipped reading M values that were in fact there.

The reader checked for 0s in the file mbox to determine if it should read m values on individual shape types. This was also done by only checking for any 0 value, while they probably meant to check for both values being 0.

But from how I read the ESRI spec, the whole checking should not be done in the first place. A file mbox of 0s is just a recommendation for files that or not Z or M type, and should have no independent meaning. Whether or not to read M values should be primarily dependent on the file type. The mbox containing 0s should not negate the fact that the file type is M or Z box, and thus contains M bytes that need to be read. Furthermore, an mbox of [0,0] is perfectly possible, eg in a M or Z-type shapefile containing only M values of 0.