Rewrote what seems to have been incorrect writing of z/m values
The code for writing z and m values seems to have been a little off in some places, so fixed how I think it was intended. Might explain some of the issues that have been raised on this in the past.

Also more consistently applied the logic of autosetting values in the edge cases when m and z values are missing, to nodata and 0, respectively.

Also documented how the z/m writing sections works in comments.