Cleared up some misunderstandings of the structure of m/z types
Internally, shapefiles treat points as a sequence of x,y values, while z values and m values are treated as their own separate sequences.

PyShp however, seems to have added an alternative convenient way of specifying m and z values. In addition to allowing setting them as separate sequences in the z and m attributes of the shape object, PyShp allows grouping them together with their respective xy coordinates using higher 3 or 4 dimensional coordinates.

In the implementation however, there seems to have been a mixup regarding the structure of z and m type shapes. PyShp treated them as incremental, that m was always just a fourth add on to the third z coordinate.

But in the ESRI spec, M type shapes only have dimensions x,y,m. Z type shapes have x,y,z,m(optional). Therefore, PyShp coordinates for m types should not contain x,y,z,m, it should only contain x,y,m.

For instance, the point() method added it as x,y,z for z-types (omitting the m value), and as x,y,z,m for m-types (including a z that shouldn't be there). I think this conceptualization may have led to incorrect files.

This commit switches this to the structure as laid out in the ESRI specs. Doings so brings it more in line with the shapefile data structure and avoids storing an unnecessary z value when it should only be storing m values, and not storing m values when it actually should.