Fix bug in hole in polygon algo
In trying to find a sample point, we iterate the ring's triplet coordinates and add add the second coord from the end to allow checking the last triplet. Previously the second coordinate was added by changing the main coords list in-place, but that messes with the point-in-polygon algorithm. Fixed by only iterating through the points instead of changing the coords list.
Fixes #205.