build.yml: Change python-versions to leave off patch version.
This makes sure that the build uses the latest patch version for each
python major.minor. Keep the pre-release for 3.13 explicit as I don't
know how that will work before the first 3.13 release is out.

Also make each version number be on a separate line so that adding and
removing python versions is easier to see in diffs, e.g., like this:

https://github.com/google/earthengine-api/blob/e44bb391af90c9b7dfde79d95b55116e79d0d6bf/.github/workflows/ci-tests.yml#L17