Delete src/py.typed
When this gets shipped into site-packages, becuase PyShp is currently a signle file libary, py.typed gets put in the root dir, flagging all other libraries the user has installed as typed, even if they aren't.  

Compliance with the typing system standard requires refactoring into a package.  This is a temporary fix for the above until a decision is made on if/how to do that.

Thanks for pointing this out Thomas @dl1jbe